% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tbl.R
\name{datatable2}
\alias{datatable2}
\title{datatable with child rows}
\usage{
datatable2(x, vars = NULL, opts = NULL, ...)
}
\arguments{
\item{vars}{столбцы для child rows}

\item{...}{}
}
\value{

}
\description{
datatable with child rows
}
