\docType{package}
\name{FI}
\alias{FI}
\alias{FI-package}
\title{Forest inventory equations and calculations with FI}
\description{
  That package was created because I can't find any free
  solution for these basic calculations also will serve as
  learning experience to create packages in R.
}
\details{
  Here I implement the three basic volume equations
  (Smalian, Newton, Huber) and form and stacking factors.
  From these you can gather many information about a region
  log volume (mean and standard deviation). Later a new
  package will provide volumetric equations adjust.
}
\seealso{
  \code{\link{volume}}
}

