% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fi_single.R
\name{fi.single}
\alias{fi.single}
\title{Fragility Index in Single Arm Design with Time to Event Endpoints}
\usage{
fi.single(data, median_benchmark, target_probability = 0.7)
}
\arguments{
\item{data}{a data matrix with at least first two columns -- time and 
censor indicator takes two values -- 1 if an event and
0 otherwise. The name of the columns must be "time" and
"status" respectively}

\item{median_benchmark}{the desired median survival beyond which the outcome 
is desired.}

\item{target_probability}{the probability threshold which is desired. 
generally very high value, default is 0.7.}
}
\value{
\item{fi}{Fragility Index}
\item{out}{An out put with time, indicator that the censored individual had been 
           converted to an event, and the corresponding posterior probability}
}
\description{
Fragility Index in Single Arm Design with Time to Event Endpoints
}
\details{
In an Oncology single arm trial often the decision is to invest 
         more on the compound if posterior probability that the median
         survival (or progression free survival) is more than median_probability
         is higher than the target probability. Here the posterior probability
         of the median survival is computed assuming the data follows an
         Exponential distribution.
}
\examples{
require(survival)
data(lung)
n <- 30 # sample size for hypothtical expansion cohort
set.seed(100) # fix the seed to sample from the full data
sample_id <- sample(x = 1:nrow(lung), size = n, replace = FALSE, prob = NULL)
lung_sample <- lung[sample_id, ]
lung_sample$timem <- lung_sample$time/30. # time in months
lung_sample$time <- lung_sample$timem
lung_sample$status <- lung_sample$status - 1 # the censored indicator is 1/2

# Kaplan Meier Plot
require(tidyverse)
ggsurvfit::survfit2(Surv(timem, status) ~ 1, data = lung_sample) \%>\% 
ggsurvfit::ggsurvfit() +
labs(
x = "Months",
y = "Overall survival probability"
) + 
ggsurvfit::add_confidence_interval() +
ggsurvfit::add_risktable()

survfit(Surv(timem, status) ~ 1, data = lung_sample) 
# median of data

fi.single(data = lung_sample, median_benchmark = 7)
# Fragility index is 5 because when the 6th censored observation
# is switched to an event, the posterior probability becomes lower
# than the thersold 0.7.

}
\references{
Maity, A. K. and Basu, C. (2024) Fragility Index in Single Arm Time to Event Trial
}
