% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz.R
\name{layer_cohort}
\alias{layer_cohort}
\title{Returns ggproto objects in a list and that list plays nicely with +.
data \%>\%
ggplot(aes(blah blah)) +
geom_something_else() +
layer_cohort(data = other_data, tweak1, tweak2)}
\usage{
layer_cohort(
  data,
  time = "MONTHS_SUBSCRIBED",
  metric = "IND_ACTIVE",
  group = "COHORT",
  ...
)
}
\arguments{
\item{...}{}
}
\value{

}
\description{
Returns ggproto objects in a list and that list plays nicely with +.
data \%>\%
ggplot(aes(blah blah)) +
geom_something_else() +
layer_cohort(data = other_data, tweak1, tweak2)
}
