\name{sf}
\alias{sf}
\title{Stacking factor}
\usage{
  sf(volume, height, length, depth)
}
\arguments{
  \item{volume}{real volume of the logs}

  \item{height}{height of the stack (in meters)}

  \item{length}{length of the stack (in meters)}

  \item{depth}{depth of the stack (in meters)}
}
\value{
  stacking factor ranging from 0-1
}
\description{
  Ratio between solid cubic meters of wood per stere cubic
  meter of stacked up wood.
}
\references{
  \url{www.eucalyptus.com.br/capitulos/ENG07.pdf}
}

