% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tbl.R
\name{showDT}
\alias{showDT}
\title{Заполнить таблицу DT данными в колонках}
\usage{
showDT(tbl, dt_caption)
}
\arguments{
\item{tbl}{таблица с колонками to_date,forecast,diff}

\item{dt_caption}{}
}
\value{

}
\description{
Заполнить таблицу DT данными в колонках
}
\examples{
 # dt |> showDT('')
}
