% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tbl.R
\name{showDT_avg}
\alias{showDT_avg}
\title{Fill table DT with data in columns with given names}
\usage{
showDT_avg(tbl, dt_caption)
}
\arguments{
\item{dt_caption}{}
}
\value{

}
\description{
Fill table DT with data in columns with given names
}
\examples{
# dt |> showDT_avg('')
}
