% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz.R
\name{validate_mypackage_data}
\alias{validate_mypackage_data}
\title{catch potential errors caused by missing keywords}
\usage{
validate_mypackage_data(vbl_names)
}
\arguments{
\item{vbl_names}{}
}
\value{

}
\description{
catch potential errors caused by missing keywords
}
