% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz.R
\name{viz_bar}
\alias{viz_bar}
\title{Статичный график двух и более переменных (столбики ggplot)}
\usage{
viz_bar(dt, x, y, z)
}
\arguments{
\item{z}{}
}
\value{

}
\description{
Статичный график двух и более переменных (столбики ggplot)
}
