% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz.R
\name{viz_cohort}
\alias{viz_cohort}
\title{basic visualization function that wraps ggplot2 code
but allows users to input their preferred x-axis, y-axis and grouping variables to draw cohort curves.}
\usage{
viz_cohort(
  data,
  metric = "IND_ACTIVE",
  time = "MONTHS_SUBSCRIBED",
  group = "COHORT",
  ...
)
}
\arguments{
\item{group}{}
}
\value{

}
\description{
basic visualization function that wraps ggplot2 code
but allows users to input their preferred x-axis, y-axis and grouping variables to draw cohort curves.
}
\examples{
#viz_cohort("my_org_data")
          #viz_cohort("my_org_data", color = COHORT, linetype = COHORT)

}
