% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz.R
\name{viz_lines}
\alias{viz_lines}
\title{Статичный график двух линий ggplot2. Например: Органик/Неорганик.}
\usage{
viz_lines(dt, x, y, z, title = "", title_x = "")
}
\arguments{
\item{title_x}{}
}
\value{

}
\description{
Статичный график двух линий ggplot2. Например: Органик/Неорганик.
}
