% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz.R
\name{viz_lines_bands_js}
\alias{viz_lines_bands_js}
\title{График факта, прогноза и прогнозного интервала}
\usage{
viz_lines_bands_js(dt, x, y1, y2, lwr, upr, title = "", title_x = "")
}
\arguments{
\item{title_x}{}
}
\value{

}
\description{
График факта, прогноза и прогнозного интервала
}
