% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz.R
\name{viz_lines_js}
\alias{viz_lines_js}
\title{Интерактивный график двух линий Органик/Неорганик, e_charts, JavaScript}
\usage{
viz_lines_js(dt, x, y, title = "", title_x = "")
}
\arguments{
\item{dt}{\itemize{
\item табличка с легендой из group_by
}}

\item{x}{\itemize{
\item название столбца дат
}}

\item{y}{\itemize{
\item название столбца выручка
}}

\item{title_x}{}
}
\value{

}
\description{
Интерактивный график двух линий Органик/Неорганик, e_charts, JavaScript
}
