% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz.R
\name{viz_lines_wide4_js}
\alias{viz_lines_wide4_js}
\title{Vizualisation with JavaScript: 4 lines from wide table with 4 columns for each measure}
\usage{
viz_lines_wide4_js(dt, x, y1, y2, y3, y4, title = "", title_x = "")
}
\arguments{
\item{x}{X axis}

\item{y1}{column name for first measure}

\item{title}{title for Vizualisation}

\item{title_x}{title for X axis}
}
\value{

}
\description{
Vizualisation with JavaScript: 4 lines from wide table with 4 columns for each measure
}
