% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tbl.R
\name{showDT_avg}
\alias{showDT_avg}
\title{Заполнить таблицу DT данными в колонках с названиями ...}
\usage{
showDT_avg(tbl, dt_caption)
}
\arguments{
\item{dt_caption}{}
}
\value{

}
\description{
Заполнить таблицу DT данными в колонках с названиями ...
}
\examples{
# dt |> showDT_avg('')
}
